refactor: rename cal oAuth to simply oauth
