+++ apps/api/v2/src/modules/oauth/oauth-client.module.ts
@@ -1,6 +1,6 @@
-import { OAuthClientController } from "@/modules/oAuth/cal/oauth-client.controller";
+import { OAuthClientController } from "@/modules/oauth/oauth-client.controller";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
-import { OAuthClientRepository } from "@/modules/repositories/oAuth/cal/oauth-client-repository.service";
+import { OAuthClientRepository } from "@/modules/repositories/oauth/oauth-client-repository.service";
 import { Module } from "@nestjs/common";
 
 @Module({
